+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Documentation of release versions of `nacc-attribute-deriver`
 * Adds logic to correlate standalone forms (COVID, CLS) to the closest UDS visit - assumes all UDS visits have been curated already
     * Original `visitdate` moved to `c19visitdate` and `clsvisitdate`, respectively
 * Standardizes all UDS `FRMDATEX` variables to `YYYY-MM-DD` format
+* Enforces ranges for several variables
 * Updates COVID `C19TxYR` variables to fix 88 to 8888 and 99 to 9999
 * Fixes MQT's `SEX MAPPING` value 9 from "Don't know" to "Unknown"